MetaBot V2/requirements.txt
python-telegram-bot==20.3
pymysql==1.0.2
psycopg2==2.9.7
MetaTrader5==5.0.37
